+++ 3D/src/renderer/Lights.cpp
@@ -14,7 +14,7 @@ void PointLight::updateRadius()
 
 inline float PointLight::calculateRadius(float linear, float quad)
 {
-	return (0.5 * (std::sqrtf(linear*linear + (Engine::cfg.render.minimumLightConstant * quad) - linear))) / quad;
+	return (0.5 * (std::sqrt(linear*linear + (Engine::cfg.render.minimumLightConstant * quad) - linear))) / quad;
 }
 
 void SpotLight::updateRadius()
@@ -25,7 +25,7 @@ void SpotLight::updateRadius()
 
 inline float SpotLight::calculateRadius(float linear, float quad)
 {
-	return (0.5 * (std::sqrtf(linear*linear + (Engine::cfg.render.minimumLightConstant * quad) - linear))) / (quad * 2.f);
+	return (0.5 * (std::sqrt(linear*linear + (Engine::cfg.render.minimumLightConstant * quad) - linear))) / (quad * 2.f);
 }
 
 PointLight & LightManager::addPointLight(PointLight::GPUData& data)
